requirements.txt
editdistance
Pillow
pytesseract
